move stopPropagation to global util
