+++ src/components/AudioPlayer/PlaybackControls.tsx
@@ -1,5 +1,6 @@
 import { useSelector } from 'react-redux';
 import { selectIsPlaying } from 'src/redux/slices/AudioPlayer/state';
+import { withStopPropagation } from 'src/utils/event';
 import styles from './PlaybackControls.module.scss';
 import PlayIcon from '../../../public/icons/play-circle-outline.svg';
 import ForwardIcon from '../../../public/icons/forward_10.svg';
@@ -8,7 +9,6 @@ import ReplayIcon from '../../../public/icons/replay_10.svg';
 import ShareIcon from '../../../public/icons/share.svg';
 import RepeatIcon from '../../../public/icons/ic_repeat_24px 1.svg';
 import Button, { ButtonSize } from '../dls/Button/Button';
-import { withStopPropagation } from './util';
 import { triggerPauseAudio, triggerPlayAudio, triggerSeek } from './EventTriggers';
 
 const PlaybackControls = () => {
